src/parallel_decoder_transformer/__init__.py
"""Parallel Decoder Transformer package now wired onto GPT-OSS primitives."""

from importlib import import_module
from pathlib import Path
from typing import Any

from dotenv import load_dotenv

_PROJECT_ROOT = Path(__file__).resolve().parents[2]
load_dotenv(_PROJECT_ROOT / ".env", override=False)

__all__ = ("config", "data", "inference", "integration", "models", "utils")


def __getattr__(name: str) -> Any:
    if name in __all__:
        module = import_module(f"{__name__}.{name}")
        globals()[name] = module
        return module
    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
